Add files via upload
Problem Statement
In the banking industry, banks do carry out a plethora of marketing campaigns to introduce new products. Sometimes, the customers don't subscribe to the offers being marketed to them and this takes a toll on the bank with regard to funds being used to market such a product.

Based on this, the need to actually know if customers will signup for a product being released and also target those exact customers is very important.